Completely reworked S-number convergence criterion
Now the extrema-finding loop counts the number of zero-crossings, and
returns it. Also, the S-number convergence criterion now takes the
number of zero crossings into account, instead of only checking that
maxima are positive and minima negative. This hopefully makes the
S-number criterion more in line with how it was originally defined by
Huang and others.

This commit should also fix a problem in the definition of the S-number
criterion: thanks to the extrapolation of extrema at the ends, the
merging of flat slopes into a single extremum, and finite precision
arithmetic, the sifting process could end up in an endless loop, where
consequent siftings change the number of extrema or zero crossings by
one. Now, the number of extrema and zero crossings are considered
"stable" to the S-number criterion if one of them changes by one. This
should not affect the S-number criterion in normal circumstances.